requirements.txt
streamlit
langchain
langchain-groq
langchain-community
psycopg2-binary